refactor(proptostyle): change file location to ease unit tests and object handling refactoring
